Submission as a List of Submittables.
